feat: add DTOs, validation and exception handler

+++ src/main/java/com/mateus/ecommerce/service/ProductService.java
@@ -1,6 +1,9 @@
 package com.mateus.ecommerce.service;
 
+import com.mateus.ecommerce.dto.ProductRequest;
+import com.mateus.ecommerce.dto.ProductResponse;
 import com.mateus.ecommerce.entity.Product;
+import com.mateus.ecommerce.mapper.ProductMapper;
 import com.mateus.ecommerce.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
@@ -10,24 +13,61 @@ import java.util.List;
 public class ProductService {
 
     private final ProductRepository repository;
+    private final ProductMapper mapper;
 
-    public ProductService(ProductRepository repository) {
+    public ProductService(
+            ProductRepository repository,
+            ProductMapper mapper) {
         this.repository = repository;
+        this.mapper = mapper;
     }
 
-    public List<Product> listarTodos() {
-        return repository.findAll();
+    public List<ProductResponse> listarTodos() {
+        return repository.findAll()
+                .stream()
+                .map(mapper::toResponse)
+                .toList();
     }
 
-    public Product salvar(Product product) {
-        return repository.save(product);
+    public ProductResponse buscarPorId(Long id) {
+        Product product = repository.findById(id)
+                .orElse(null);
+
+        if (product == null) {
+            return null;
+        }
+
+        return mapper.toResponse(product);
     }
 
-    public Product buscarPorId(Long id) {
-        return repository.findById(id).orElse(null);
+    public ProductResponse salvar(ProductRequest request) {
+        Product product = mapper.toEntity(request);
+        Product produtoSalvo = repository.save(product);
+
+        return mapper.toResponse(produtoSalvo);
     }
 
-    public void excluir(Long id) {
+    public ProductResponse atualizar(Long id, ProductRequest request) {
+        Product product = repository.findById(id)
+                .orElse(null);
+
+        if (product == null) {
+            return null;
+        }
+
+        mapper.updateEntity(request, product);
+
+        Product produtoAtualizado = repository.save(product);
+
+        return mapper.toResponse(produtoAtualizado);
+    }
+
+    public boolean excluir(Long id) {
+        if (!repository.existsById(id)) {
+            return false;
+        }
+
         repository.deleteById(id);
+        return true;
     }
 }
